requirements.txt
boto==2.43.0
botocore==1.7.19
boto3==1.4.7
Django>=1.11.18
django-braces==1.13.0
django-ckeditor==5.6.1
django-crispy-forms>=1.7.0
django-debug-toolbar
django-s3direct==0.4.4
django-storages==1.5.1
docb
envs==1.2.6
Unipath==1.1
zappa==0.47.1
git+https://github.com/metametricsinc/django-warrant@feature/warrant-lite
